cli: remove the "debug kv" commands
Removed the use of `debug kv scan` from the reference tests. Those scans
were not doing what the author thought they were doing as `kv scan` with
no arguments only scans the "user data
span" (i.e. [/System/Max,/Table/Min]).

Fixes #14842
