+++ pkg/cli/cli_test.go
@@ -319,9 +319,6 @@ communicate with a secure cluster\).
 	}{
 		// Error returned from GRPC to internal/client (which has to pass it
 		// up the stack as a roachpb.NewError(roachpb.NewSendError(.)).
-		{`debug kv inc a`, styled(
-			`increment failed: failed to send RPC: rpc error: code = Unavailable desc = grpc: the connection is unavailable`),
-		},
 		// Error returned directly from GRPC.
 		{`quit`, styled(
 			`Failed to connect to the node: error sending drain request: rpc error: code = Unavailable desc = grpc: the connection is unavailable`),
@@ -341,9 +338,6 @@ communicate with a secure cluster\).
 		{`zone ls`, styled(
 			`dial tcp .*: .* refused`),
 		},
-		// A real error before we hit any networking ones.
-		{`debug kv scan b a`, `start key must be smaller than end key`},
-		{`debug kv inc`, `invalid arguments`},
 	} {
 		t.Run(test.cmd, func(t *testing.T) {
 			out, err := c.RunWithCapture(test.cmd)
@@ -359,141 +353,14 @@ communicate with a secure cluster\).
 	}
 }
 
-func Example_basic() {
-	c := newCLITest(cliTestParams{})
-	defer c.cleanup()
-
-	c.Run("debug kv put basic1 1 basic2 2 basic3 3")
-	c.Run("debug kv scan")
-	c.Run("debug kv revscan")
-	c.Run("debug kv del basic1 basic3")
-	c.Run("debug kv get basic1")
-	c.Run("debug kv get basic2")
-	c.Run("debug kv inc basic3 1")
-	c.Run("debug kv inc basic3 10")
-	c.Run("debug kv inc basic3 100")
-	c.Run("debug kv inc basic3 -- -60")
-	c.Run("debug kv inc basic3 -- -9")
-	c.Run("debug kv scan")
-	c.Run("debug kv revscan")
-	c.Run("debug kv inc basic3 b")
-
-	// Output:
-	// debug kv put basic1 1 basic2 2 basic3 3
-	// debug kv scan
-	// "basic1"	"1"
-	// "basic2"	"2"
-	// "basic3"	"3"
-	// 3 result(s)
-	// debug kv revscan
-	// "basic3"	"3"
-	// "basic2"	"2"
-	// "basic1"	"1"
-	// 3 result(s)
-	// debug kv del basic1 basic3
-	// debug kv get basic1
-	// "basic1" not found
-	// debug kv get basic2
-	// "2"
-	// debug kv inc basic3 1
-	// 1
-	// debug kv inc basic3 10
-	// 11
-	// debug kv inc basic3 100
-	// 111
-	// debug kv inc basic3 -- -60
-	// 51
-	// debug kv inc basic3 -- -9
-	// 42
-	// debug kv scan
-	// "basic2"	"2"
-	// "basic3"	42
-	// 2 result(s)
-	// debug kv revscan
-	// "basic3"	42
-	// "basic2"	"2"
-	// 2 result(s)
-	// debug kv inc basic3 b
-	// invalid increment: strconv.ParseInt: parsing "b": invalid syntax
-}
-
-func Example_quoted() {
-	c := newCLITest(cliTestParams{})
-	defer c.cleanup()
-
-	c.Run(`debug kv put quoted\x00 日本語`)                                  // UTF-8 input text
-	c.Run(`debug kv put quoted\x01 \u65e5\u672c\u8a9e`)                   // explicit Unicode code points
-	c.Run(`debug kv put quoted\x02 \U000065e5\U0000672c\U00008a9e`)       // explicit Unicode code points
-	c.Run(`debug kv put quoted\x03 \xe6\x97\xa5\xe6\x9c\xac\xe8\xaa\x9e`) // explicit UTF-8 bytes
-	c.Run(`debug kv scan`)
-	c.Run(`debug kv get quoted\x00`)
-	c.Run(`debug kv del quoted\x00`)
-	c.Run(`debug kv inc quoted\x04`)
-	c.Run(`debug kv get quoted\x04`)
-
-	// Output:
-	// debug kv put quoted\x00 日本語
-	// debug kv put quoted\x01 \u65e5\u672c\u8a9e
-	// debug kv put quoted\x02 \U000065e5\U0000672c\U00008a9e
-	// debug kv put quoted\x03 \xe6\x97\xa5\xe6\x9c\xac\xe8\xaa\x9e
-	// debug kv scan
-	// "quoted\x00"	"日本語"
-	// "quoted\x01"	"日本語"
-	// "quoted\x02"	"日本語"
-	// "quoted\x03"	"日本語"
-	// 4 result(s)
-	// debug kv get quoted\x00
-	// "日本語"
-	// debug kv del quoted\x00
-	// debug kv inc quoted\x04
-	// 1
-	// debug kv get quoted\x04
-	// 1
-}
-
-func Example_insecure() {
-	c := newCLITest(cliTestParams{insecure: true})
-	defer c.cleanup()
-
-	c.Run("debug kv put insecure1 1 insecure2 2")
-	c.Run("debug kv scan")
-
-	// Output:
-	// debug kv put insecure1 1 insecure2 2
-	// debug kv scan
-	// "insecure1"	"1"
-	// "insecure2"	"2"
-	// 2 result(s)
-}
-
 func Example_ranges() {
 	c := newCLITest(cliTestParams{})
 	defer c.cleanup()
 
-	c.Run("debug kv put ranges1 1 ranges2 2 ranges3 3 ranges4 4")
-	c.Run("debug kv scan")
-	c.Run("debug kv revscan")
 	c.Run("debug range split ranges3")
 	c.Run("debug range ls")
-	c.Run("debug kv scan")
-	c.Run("debug kv revscan")
-	c.Run("debug kv delrange ranges1 ranges3")
-	c.Run("debug kv scan")
 
 	// Output:
-	// debug kv put ranges1 1 ranges2 2 ranges3 3 ranges4 4
-	// debug kv scan
-	// "ranges1"	"1"
-	// "ranges2"	"2"
-	// "ranges3"	"3"
-	// "ranges4"	"4"
-	// 4 result(s)
-	// debug kv revscan
-	// "ranges4"	"4"
-	// "ranges3"	"3"
-	// "ranges2"	"2"
-	// "ranges1"	"1"
-	// 4 result(s)
 	// debug range split ranges3
 	// debug range ls
 	// /Min-/System/"" [1]
@@ -519,23 +386,6 @@ func Example_ranges() {
 	// /Table/15-/Max [10]
 	// 	0: node-id=1 store-id=1
 	// 11 result(s)
-	// debug kv scan
-	// "ranges1"	"1"
-	// "ranges2"	"2"
-	// "ranges3"	"3"
-	// "ranges4"	"4"
-	// 4 result(s)
-	// debug kv revscan
-	// "ranges4"	"4"
-	// "ranges3"	"3"
-	// "ranges2"	"2"
-	// "ranges1"	"1"
-	// 4 result(s)
-	// debug kv delrange ranges1 ranges3
-	// debug kv scan
-	// "ranges3"	"3"
-	// "ranges4"	"4"
-	// 2 result(s)
 }
 
 func Example_logging() {
@@ -576,57 +426,15 @@ func Example_logging() {
 	// 1
 }
 
-func Example_cput() {
-	c := newCLITest(cliTestParams{})
-	defer c.cleanup()
-
-	c.Run("debug kv put cput1 1 cput2 2 cput3 3 cput4 4")
-	c.Run("debug kv scan")
-	c.Run("debug kv cput cput5 5")
-	c.Run("debug kv cput cput2 3 2")
-	c.Run("debug kv scan")
-
-	// Output:
-	// debug kv put cput1 1 cput2 2 cput3 3 cput4 4
-	// debug kv scan
-	// "cput1"	"1"
-	// "cput2"	"2"
-	// "cput3"	"3"
-	// "cput4"	"4"
-	// 4 result(s)
-	// debug kv cput cput5 5
-	// debug kv cput cput2 3 2
-	// debug kv scan
-	// "cput1"	"1"
-	// "cput2"	"3"
-	// "cput3"	"3"
-	// "cput4"	"4"
-	// "cput5"	"5"
-	// 5 result(s)
-}
-
 func Example_max_results() {
 	c := newCLITest(cliTestParams{})
 	defer c.cleanup()
 
-	c.Run("debug kv put max_results1 1 max_results2 2 max_results3 3 max_results4 4")
-	c.Run("debug kv scan --max-results=3")
-	c.Run("debug kv revscan --max-results=2")
 	c.Run("debug range split max_results3")
 	c.Run("debug range split max_results4")
 	c.Run("debug range ls --max-results=5")
 
 	// Output:
-	// debug kv put max_results1 1 max_results2 2 max_results3 3 max_results4 4
-	// debug kv scan --max-results=3
-	// "max_results1"	"1"
-	// "max_results2"	"2"
-	// "max_results3"	"3"
-	// 3 result(s)
-	// debug kv revscan --max-results=2
-	// "max_results4"	"4"
-	// "max_results3"	"3"
-	// 2 result(s)
 	// debug range split max_results3
 	// debug range split max_results4
 	// debug range ls --max-results=5
