+++ pkg/cli/context.go
@@ -18,12 +18,14 @@ package cli
 
 import (
 	"fmt"
+	"strconv"
 	"strings"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine"
+	"github.com/pkg/errors"
 )
 
 // statementsValue is an implementation of pflag.Value that appends any
@@ -192,6 +194,16 @@ func parseKeyType(value string) (keyType, error) {
 	return 0, fmt.Errorf("unknown key type '%s'", value)
 }
 
+// unquoteArg unquotes the provided argument using Go double-quoted
+// string literal rules.
+func unquoteArg(arg string) (string, error) {
+	s, err := strconv.Unquote(`"` + arg + `"`)
+	if err != nil {
+		return "", errors.Wrapf(err, "invalid argument %q", arg)
+	}
+	return s, nil
+}
+
 type mvccKey engine.MVCCKey
 
 func (k *mvccKey) String() string {
@@ -215,7 +227,7 @@ func (k *mvccKey) Set(value string) error {
 
 	switch typ {
 	case raw:
-		unquoted, err := unquoteArg(keyStr, false)
+		unquoted, err := unquoteArg(keyStr)
 		if err != nil {
 			return err
 		}
