+++ pkg/cli/start.go
@@ -583,6 +583,17 @@ func runStart(cmd *cobra.Command, args []string) error {
 	return returnErr
 }
 
+func addrWithDefaultHost(addr string) (string, error) {
+	host, port, err := net.SplitHostPort(baseCfg.Addr)
+	if err != nil {
+		return "", err
+	}
+	if host == "" {
+		host = "localhost"
+	}
+	return net.JoinHostPort(host, port), nil
+}
+
 func getClientGRPCConn() (*grpc.ClientConn, *hlc.Clock, *stop.Stopper, error) {
 	// 0 to disable max offset checks; this RPC context is not a member of the
 	// cluster, so there's no need to enforce that its max offset is the same
